requirement.txt
ipykernel
numpy
matplotlib
pillow
opencv-python
scipy
tensorflow
tensorflow-gpu
streamlit